[Refactor] AppCompatibleDevicesView 하위 뷰로 분리
- 앱 호환기기 영역 하위 뷰로 분리
